// apps/web/src/lib/api/generated/models/courseOutcome.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Formex
 * OpenAPI spec version: 0.0.1
 */
import type { Program } from "./program";
import type { Subject } from "./subject";

export interface CourseOutcome {
  id: string;
  institution_id: string;
  program_id: string;
  course_id: string;
  type: string;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  syllabus_scheme: string | null;
  academic_year: number;
  /** @nullable */
  created_at: string | null;
  /** @nullable */
  updated_at: string | null;
  program?: Program;
  course?: Subject;
}